// src/types/ScreenshotProtocol.ts
export enum ScreenshotProtocol {
    RPIC = 'RPIC',     // request pic data
    APIC = 'APIC',     // answer pic data
    FPIC = 'FPIC',     // finish pic data
    RACT = 'RACT',     // request activity
    AACT = 'AACT',     // answer activity
    FACT = 'FACT',     // finish activity
    RXML = 'RXML',     // request xml
    AXML = 'AXML',     // answer xml
    FXML = 'FXML',     // finish xml
    RSER = 'RSER',     // request service
    ASER = 'ASER',     // answer service
    FSER = 'FSER',     // finish service
    RHIE = 'RHIE',     // request activity
    AHIE = 'AHIE',     // answer activity
    FHIE = 'FHIE',     // finish activity
}